# packages/contracts/openapi/smartai-core-v1.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "知聘招聘智能体 Core API",
    "version": "1.0.0",
    "summary": "招聘智能体生产试点 MVP 的权威 HTTP 契约",
    "description": "覆盖 ATS 嵌入会话、招聘任务、岗位方案、人才匹配、候选名单确认、在线面试、综合评价、人工门禁、知识、审计和连接器能力。服务端状态、权限、版本和人工门禁是唯一事实源；客户端不得自行推进关键业务状态。"
  },
  "servers": [
    {
      "url": "/",
      "description": "当前部署环境"
    }
  ],
...
  "security": [
    { "BearerAuth": [] }
  ],
  "paths": {
...
    "/api/core/v1/requirement-drafts": {
      "post": {
        "tags": ["RequirementDrafts"],
        "summary": "创建需求草案",
        "description": "草案不是招聘任务；只有人工确认并执行 convert 后才创建正式任务。",
        "operationId": "createRequirementDraft",
        "parameters": [
          { "$ref": "#/components/parameters/RequestId" },
          { "$ref": "#/components/parameters/CorrelationId" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/RequirementDraftCreateRequest" } }
          }
        },
        "responses": {
          "201": {
            "description": "草案已创建",
            "headers": {
              "ETag": { "$ref": "#/components/headers/ETag" },
              "X-SmartAI-Input-Hash": { "$ref": "#/components/headers/SmartAIInputHash" },
              "Idempotency-Replayed": { "$ref": "#/components/headers/IdempotencyReplayed" }
            },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RequirementDraftResponse" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/VersionConflict" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalServerError" },
          "503": { "$ref": "#/components/responses/ServiceUnavailable" },
          "504": { "$ref": "#/components/responses/GatewayTimeout" },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/core/v1/requirement-drafts/{draftId}": {
      "parameters": [{ "$ref": "#/components/parameters/DraftId" }],
      "get": {
        "tags": ["RequirementDrafts"],
        "summary": "读取需求草案",
        "operationId": "getRequirementDraft",
        "parameters": [{ "$ref": "#/components/parameters/RequestId" }],
        "responses": {
          "200": {
            "description": "需求草案",
            "headers": { "ETag": { "$ref": "#/components/headers/ETag" }, "X-SmartAI-Input-Hash": { "$ref": "#/components/headers/SmartAIInputHash" } },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RequirementDraftResponse" } } }
          },
          "default": { "$ref": "#/components/responses/Error" }
        }
      },
      "patch": {
        "tags": ["RequirementDrafts"],
        "summary": "修改未转换的需求草案",
        "operationId": "updateRequirementDraft",
        "parameters": [
          { "$ref": "#/components/parameters/RequestId" },
          { "$ref": "#/components/parameters/IfMatch" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/merge-patch+json": { "schema": { "$ref": "#/components/schemas/RequirementDraftPatchRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "草案已更新",
            "headers": { "ETag": { "$ref": "#/components/headers/ETag" }, "X-SmartAI-Input-Hash": { "$ref": "#/components/headers/SmartAIInputHash" } },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RequirementDraftResponse" } } }
          },
          "409": { "$ref": "#/components/responses/VersionConflict" },
          "428": { "$ref": "#/components/responses/PreconditionRequired" },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/core/v1/requirement-drafts/{draftId}/convert": {
      "post": {
        "tags": ["RequirementDrafts", "RecruitmentTasks"],
        "summary": "人工确认草案并创建招聘任务",
        "description": "对应人工门禁 G1。确认人来自访问令牌，不接受请求体伪造。相同草案和幂等键只创建一个任务。",
        "operationId": "convertRequirementDraft",
        "x-human-gate": "CREATE_TASK",
        "parameters": [
          { "$ref": "#/components/parameters/DraftId" },
          { "$ref": "#/components/parameters/RequestId" },
          { "$ref": "#/components/parameters/CorrelationId" },
          { "$ref": "#/components/parameters/IfMatch" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RequirementDraftConvertRequest" } } }
        },
        "responses": {
          "201": {
            "description": "招聘任务已创建",
            "headers": { "ETag": { "$ref": "#/components/headers/ETag" }, "Idempotency-Replayed": { "$ref": "#/components/headers/IdempotencyReplayed" } },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RecruitmentTaskResponse" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/VersionConflict" },
          "410": { "$ref": "#/components/responses/Gone" },
          "428": { "$ref": "#/components/responses/PreconditionRequired" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalServerError" },
          "503": { "$ref": "#/components/responses/ServiceUnavailable" },
          "504": { "$ref": "#/components/responses/GatewayTimeout" },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/core/v1/recruitment-tasks": {
      "get": {
        "tags": ["RecruitmentTasks"],
        "summary": "分页查询有权访问的招聘任务",
        "operationId": "listRecruitmentTasks",
        "parameters": [
          { "$ref": "#/components/parameters/RequestId" },
          { "$ref": "#/components/parameters/Cursor" },
          { "$ref": "#/components/parameters/Limit" },
          { "name": "businessStage", "in": "query", "schema": { "$ref": "#/components/schemas/BusinessStage" } },
          { "name": "lifecycleStatus", "in": "query", "schema": { "$ref": "#/components/schemas/LifecycleStatus" } },
          { "name": "ownerUserId", "in": "query", "schema": { "$ref": "#/components/schemas/Uuid" } }
        ],
        "responses": {
          "200": {
            "description": "招聘任务页",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RecruitmentTaskPageResponse" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalServerError" },
          "503": { "$ref": "#/components/responses/ServiceUnavailable" },
          "504": { "$ref": "#/components/responses/GatewayTimeout" },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/core/v1/recruitment-tasks/{taskId}": {
      "parameters": [{ "$ref": "#/components/parameters/TaskId" }],
      "get": {
        "tags": ["RecruitmentTasks"],
        "summary": "读取招聘任务",
        "operationId": "getRecruitmentTask",
        "parameters": [{ "$ref": "#/components/parameters/RequestId" }],
        "responses": {
          "200": {
            "description": "招聘任务",
            "headers": { "ETag": { "$ref": "#/components/headers/ETag" } },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RecruitmentTaskResponse" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalServerError" },
          "503": { "$ref": "#/components/responses/ServiceUnavailable" },
          "504": { "$ref": "#/components/responses/GatewayTimeout" },
          "default": { "$ref": "#/components/responses/Error" }
        }
      },
      "patch": {
        "tags": ["RecruitmentTasks"],
        "summary": "修改任务非状态字段",
        "description": "业务阶段和生命周期只能通过领域命令或门禁推进，不接受 PATCH 直接修改。",
        "operationId": "updateRecruitmentTask",
        "parameters": [
          { "$ref": "#/components/parameters/RequestId" },
          { "$ref": "#/components/parameters/IfMatch" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/merge-patch+json": { "schema": { "$ref": "#/components/schemas/RecruitmentTaskPatchRequest" } } }
        },
        "responses": {
          "200": {
            "description": "招聘任务已更新",
            "headers": { "ETag": { "$ref": "#/components/headers/ETag" } },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RecruitmentTaskResponse" } } }
          },
          "409": { "$ref": "#/components/responses/VersionConflict" },
          "428": { "$ref": "#/components/responses/PreconditionRequired" },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "用户或 iframe 短期访问令牌；租户、用户、scope 和资源范围从已验证令牌解析。"
      },
      "ServiceOAuth": {
        "type": "oauth2",
        "description": "客户 ATS 后端的最小权限服务身份。",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "/oauth2/token",
            "scopes": { "embed:session:create": "创建用户委托的嵌入会话" }
          }
        }
      }
    },
    "parameters": {
      "RequestId": { "name": "X-Request-Id", "in": "header", "required": false, "description": "调用方生成的追踪 ID；缺失时服务端生成。", "schema": { "$ref": "#/components/schemas/Uuid" } },
      "CorrelationId": { "name": "X-Correlation-Id", "in": "header", "required": false, "description": "招聘任务、运行或宿主会话关联 ID。", "schema": { "type": "string", "minLength": 1, "maxLength": 128 } },
      "IdempotencyKey": { "name": "Idempotency-Key", "in": "header", "required": true, "description": "写请求必填。相同键和相同请求摘要返回首次结果；相同键但载荷不同返回 IDEMPOTENCY_CONFLICT。", "schema": { "type": "string", "format": "uuid" } },
      "IfMatch": { "name": "If-Match", "in": "header", "required": true, "description": "资源版本 ETag，例如 W/\"7\"。缺失返回 428，版本冲突返回 409。", "schema": { "type": "string", "pattern": "^(W/)?\\\"[1-9][0-9]*\\\"$" } },
      "Cursor": { "name": "cursor", "in": "query", "required": false, "description": "服务端返回的不透明游标。", "schema": { "type": "string", "maxLength": 512 } },
      "Limit": { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "maximum": 200, "default": 50 } },
...
      "DraftId": { "name": "draftId", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/Uuid" } },
      "TaskId": { "name": "taskId", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/Uuid" } },
...
    },
    "headers": {
      "ETag": { "description": "当前资源版本，格式 W/\"{version}\"。", "schema": { "type": "string" } },
      "SmartAIInputHash": { "description": "当前招聘需求草案的规范化确认摘要。G1 确认必须原样回传。", "schema": { "type": "string", "pattern": "^[a-f0-9]{64}$" } },
      "IdempotencyReplayed": { "description": "true 表示返回了同一幂等请求的首次结果。", "schema": { "type": "boolean" } }
    },
    "responses": {
      "Error": {
        "description": "统一错误响应",
        "headers": { "Retry-After": { "description": "可重试错误建议等待秒数。", "schema": { "type": "integer", "minimum": 0 } } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
      },
      "BadRequest": {
        "description": "请求格式、字段约束、游标或业务输入校验失败",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
      },
      "Unauthorized": {
        "description": "访问令牌、服务身份或一次性引导令牌无效",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
      },
      "Forbidden": {
        "description": "调用方已认证但缺少 scope、数据范围、资源权限或允许的宿主 Origin",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
      },
      "NotFound": {
        "description": "资源不存在、不可见或外部引用无法映射",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
      },
      "VersionConflict": {
        "description": "版本或幂等冲突",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
      },
      "Gone": {
        "description": "一次性令牌、嵌入会话、上下文解析结果或需求草案已过期且不可恢复",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
      },
      "PreconditionRequired": {
        "description": "修改版本化资源时缺少 If-Match",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
      },
      "RateLimited": {
        "description": "调用频率或租户配额已超限",
        "headers": { "Retry-After": { "description": "建议等待秒数。", "schema": { "type": "integer", "minimum": 0 } } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
      },
      "InternalServerError": {
        "description": "平台内部错误；响应和日志不得泄露令牌、个人信息或内部堆栈",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
      },
      "ServiceUnavailable": {
        "description": "平台依赖或上游服务暂时不可用",
        "headers": { "Retry-After": { "description": "建议等待秒数。", "schema": { "type": "integer", "minimum": 0 } } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
      },
      "GatewayTimeout": {
        "description": "平台等待身份、映射、连接器或 AI 上游响应超时",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
      }
    },
    "schemas": {
      "Uuid": { "type": "string", "format": "uuid" },
...
      "Version": { "type": "integer", "minimum": 1 },
      "SuccessEnvelope": {
        "type": "object",
        "required": ["requestId", "data", "meta"],
        "properties": {
          "requestId": { "$ref": "#/components/schemas/Uuid" },
          "data": {},
          "meta": {
            "oneOf": [
              { "$ref": "#/components/schemas/ResponseMeta" },
              { "$ref": "#/components/schemas/PageMeta" }
            ]
          }
        },
        "additionalProperties": false
      },
      "ResponseMeta": {
        "type": "object",
        "required": ["servedAt"],
        "properties": {
          "servedAt": { "type": "string", "format": "date-time" },
          "traceId": { "type": "string", "maxLength": 64 }
        },
        "additionalProperties": false
      },
      "PageMeta": {
        "type": "object",
        "required": ["servedAt", "limit", "hasMore"],
        "properties": {
          "servedAt": { "type": "string", "format": "date-time" },
          "traceId": { "type": "string", "maxLength": 64 },
          "limit": { "type": "integer", "minimum": 1, "maximum": 200 },
          "hasMore": { "type": "boolean" },
          "nextCursor": { "type": ["string", "null"], "maxLength": 512 }
        },
        "additionalProperties": false
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": ["requestId", "error"],
        "properties": {
          "requestId": { "$ref": "#/components/schemas/Uuid" },
          "error": { "$ref": "#/components/schemas/ApiError" }
        },
        "additionalProperties": false
      },
      "ApiError": {
        "type": "object",
        "required": ["code", "message", "retryable", "traceId"],
        "properties": {
          "code": { "$ref": "#/components/schemas/ErrorCode" },
          "message": { "type": "string", "minLength": 1, "maxLength": 500 },
          "retryable": { "type": "boolean" },
          "retryAfterSeconds": { "type": ["integer", "null"], "minimum": 0 },
          "traceId": { "type": "string", "minLength": 1, "maxLength": 64 },
          "details": { "type": "array", "items": { "$ref": "#/components/schemas/ErrorDetail" }, "maxItems": 100 }
        },
        "additionalProperties": false
      },
      "ErrorDetail": {
        "type": "object",
        "required": ["code"],
        "properties": {
          "code": { "type": "string", "maxLength": 80 },
          "field": { "type": ["string", "null"], "maxLength": 200 },
          "message": { "type": ["string", "null"], "maxLength": 500 },
          "resourceRef": { "oneOf": [{ "$ref": "#/components/schemas/ResourceRef" }, { "type": "null" }] }
        },
        "additionalProperties": false
      },
      "ErrorCode": {
        "type": "string",
        "enum": ["VALIDATION_FAILED", "AUTH_FAILED", "PERMISSION_DENIED", "RESOURCE_NOT_FOUND", "STATE_NOT_ALLOWED", "VERSION_CONFLICT", "PRECONDITION_REQUIRED", "IDEMPOTENCY_CONFLICT", "HUMAN_GATE_REQUIRED", "RATE_LIMITED", "UPSTREAM_UNAVAILABLE", "TIMEOUT", "MAPPING_FAILED", "UNSUPPORTED_CAPABILITY", "CONTEXT_EXPIRED", "CONTEXT_FORBIDDEN", "EMBED_ORIGIN_DENIED", "TOKEN_REPLAYED", "SCHEMA_VERSION_UNSUPPORTED", "INTERNAL_ERROR"]
      },
      "ExternalRef": {
        "type": "object",
        "required": ["system", "id"],
        "properties": {
          "system": { "type": "string", "pattern": "^[a-z0-9][a-z0-9._-]{1,63}$" },
          "id": { "type": "string", "minLength": 1, "maxLength": 256 },
          "objectType": { "type": ["string", "null"], "maxLength": 80 },
          "version": { "oneOf": [{ "type": "integer", "minimum": 0 }, { "type": "string", "maxLength": 128 }, { "type": "null" }] }
        },
        "additionalProperties": false
      },
      "ResourceRef": {
        "type": "object",
        "required": ["type", "id", "version"],
        "properties": {
          "type": { "type": "string", "pattern": "^[A-Z][A-Za-z0-9]{1,79}$" },
          "id": { "$ref": "#/components/schemas/Uuid" },
          "version": { "$ref": "#/components/schemas/Version" }
        },
        "additionalProperties": false
      },
      "UserRef": {
        "type": "object",
        "required": ["id", "displayName"],
        "properties": {
          "id": { "$ref": "#/components/schemas/Uuid" },
          "displayName": { "type": "string", "minLength": 1, "maxLength": 120 }
        },
        "additionalProperties": false
      },
...
      "RequirementDraftStatus": { "type": "string", "enum": ["DRAFT", "READY", "CONVERTED", "ABANDONED", "EXPIRED"] },
      "RequirementField": {
        "type": "object",
        "required": ["value", "confidence", "source", "needsConfirmation"],
        "properties": {
          "value": {},
          "confidence": { "type": "number", "minimum": 0, "maximum": 1 },
          "source": { "type": "string", "enum": ["USER", "DEFAULT", "AI", "IMPORTED"] },
          "needsConfirmation": { "type": "boolean" },
          "evidence": { "type": ["string", "null"], "maxLength": 1000 }
        },
        "additionalProperties": false
      },
      "RequirementFields": {
        "type": "object",
        "properties": {
          "positionName": { "$ref": "#/components/schemas/RequirementField" },
          "organizationRef": { "$ref": "#/components/schemas/RequirementField" },
          "locations": { "$ref": "#/components/schemas/RequirementField" },
          "headcount": { "$ref": "#/components/schemas/RequirementField" },
          "recruitmentType": { "$ref": "#/components/schemas/RequirementField" },
          "priority": { "$ref": "#/components/schemas/RequirementField" },
          "targetDate": { "$ref": "#/components/schemas/RequirementField" },
          "coreRequirements": { "$ref": "#/components/schemas/RequirementField" },
          "knowledgeScope": { "$ref": "#/components/schemas/RequirementField" }
        },
        "additionalProperties": false
      },
      "RequirementDraft": {
        "type": "object",
        "required": ["id", "status", "version", "rawInput", "fields", "createdBy", "createdAt", "updatedAt", "expiresAt"],
        "properties": {
          "id": { "$ref": "#/components/schemas/Uuid" },
          "status": { "$ref": "#/components/schemas/RequirementDraftStatus" },
          "version": { "$ref": "#/components/schemas/Version" },
          "rawInput": { "type": "string", "minLength": 1, "maxLength": 10000, "description": "按权限返回并在服务端加密存储。" },
          "fields": { "$ref": "#/components/schemas/RequirementFields" },
          "sourceJobRef": { "oneOf": [{ "$ref": "#/components/schemas/ExternalRef" }, { "type": "null" }] },
          "hostContextHash": { "type": ["string", "null"], "pattern": "^[a-f0-9]{64}$" },
          "createdBy": { "$ref": "#/components/schemas/UserRef" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "expiresAt": { "type": "string", "format": "date-time" },
          "convertedTaskRef": { "oneOf": [{ "$ref": "#/components/schemas/ResourceRef" }, { "type": "null" }] }
        },
        "additionalProperties": false
      },
      "RequirementDraftCreateRequest": {
        "type": "object",
        "required": ["input"],
        "properties": {
          "input": { "type": "string", "minLength": 10, "maxLength": 10000 },
          "sourceJobRef": { "oneOf": [{ "$ref": "#/components/schemas/ExternalRef" }, { "type": "null" }] },
          "hostContextHash": { "type": ["string", "null"], "pattern": "^[a-f0-9]{64}$" },
          "locale": { "type": "string", "default": "zh-CN" }
        },
        "additionalProperties": false
      },
      "RequirementDraftPatchRequest": {
        "type": "object",
        "properties": {
          "rawInput": { "type": "string", "minLength": 10, "maxLength": 10000 },
          "fields": { "$ref": "#/components/schemas/RequirementFields" }
        },
        "minProperties": 1,
        "additionalProperties": false
      },
      "RequirementDraftConvertRequest": {
        "type": "object",
        "required": ["confirmation", "ownerUserId"],
        "properties": {
          "confirmation": { "$ref": "#/components/schemas/HumanConfirmation" },
          "ownerUserId": { "$ref": "#/components/schemas/Uuid" },
          "hiringManagerUserId": { "oneOf": [{ "$ref": "#/components/schemas/Uuid" }, { "type": "null" }] },
          "participantUserIds": { "type": "array", "items": { "$ref": "#/components/schemas/Uuid" }, "uniqueItems": true, "maxItems": 100 }
        },
        "additionalProperties": false
      },
      "HumanConfirmation": {
        "type": "object",
        "required": ["confirmed", "inputHash"],
        "properties": {
          "confirmed": { "type": "boolean", "const": true },
          "inputHash": { "type": "string", "pattern": "^[a-f0-9]{64}$" },
          "comment": { "type": ["string", "null"], "maxLength": 2000 }
        },
        "additionalProperties": false
      },
      "RequirementDraftResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/SuccessEnvelope" },
          { "properties": { "data": { "$ref": "#/components/schemas/RequirementDraft" } } }
        ]
      },
      "BusinessStage": { "type": "string", "enum": ["ROLE_PLAN", "TALENT_SEARCH", "CANDIDATE_CONFIRMATION", "ONLINE_INTERVIEW", "COMPREHENSIVE_EVALUATION", "COMPLETED"] },
      "LifecycleStatus": { "type": "string", "enum": ["ACTIVE", "PAUSED", "CANCELLED", "ARCHIVED"] },
      "ExecutionStatus": { "type": "string", "enum": ["IDLE", "QUEUED", "RUNNING", "WAITING_HUMAN", "SUCCEEDED", "FAILED_RETRYABLE", "FAILED_FINAL", "CANCEL_REQUESTED", "CANCELLED"] },
      "RecruitmentTask": {
        "type": "object",
        "required": ["id", "taskNo", "title", "positionName", "organizationRef", "owner", "recruitmentType", "headcount", "locations", "priority", "businessStage", "lifecycleStatus", "executionStatus", "version", "creationCheckpointRef", "createdAt", "updatedAt"],
        "properties": {
          "id": { "$ref": "#/components/schemas/Uuid" },
          "taskNo": { "type": "string", "pattern": "^[A-Z0-9-]{6,40}$" },
          "title": { "type": "string", "minLength": 1, "maxLength": 200 },
          "positionName": { "type": "string", "minLength": 1, "maxLength": 160 },
          "organizationRef": { "$ref": "#/components/schemas/ResourceRef" },
          "owner": { "$ref": "#/components/schemas/UserRef" },
          "hiringManager": { "oneOf": [{ "$ref": "#/components/schemas/UserRef" }, { "type": "null" }] },
          "participants": { "type": "array", "items": { "$ref": "#/components/schemas/UserRef" }, "maxItems": 100 },
          "recruitmentType": { "type": "string", "enum": ["SOCIAL", "CAMPUS", "INTERNAL"] },
          "headcount": { "type": "integer", "minimum": 1, "maximum": 10000 },
          "locations": { "type": "array", "items": { "type": "string", "minLength": 1, "maxLength": 100 }, "minItems": 1, "uniqueItems": true },
          "priority": { "type": "string", "enum": ["LOW", "NORMAL", "HIGH", "URGENT"] },
          "targetDate": { "type": ["string", "null"], "format": "date" },
          "businessStage": { "$ref": "#/components/schemas/BusinessStage" },
          "lifecycleStatus": { "$ref": "#/components/schemas/LifecycleStatus" },
          "executionStatus": { "$ref": "#/components/schemas/ExecutionStatus" },
          "version": { "$ref": "#/components/schemas/Version" },
          "creationCheckpointRef": { "$ref": "#/components/schemas/ResourceRef" },
          "currentPlanVersionRef": { "oneOf": [{ "$ref": "#/components/schemas/ResourceRef" }, { "type": "null" }] },
          "sourceJobRef": { "oneOf": [{ "$ref": "#/components/schemas/ExternalRef" }, { "type": "null" }] },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "additionalProperties": false
      },
      "RecruitmentTaskPatchRequest": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "minLength": 1, "maxLength": 200 },
          "ownerUserId": { "$ref": "#/components/schemas/Uuid" },
          "hiringManagerUserId": { "oneOf": [{ "$ref": "#/components/schemas/Uuid" }, { "type": "null" }] },
          "participantUserIds": { "type": "array", "items": { "$ref": "#/components/schemas/Uuid" }, "uniqueItems": true, "maxItems": 100 },
          "priority": { "type": "string", "enum": ["LOW", "NORMAL", "HIGH", "URGENT"] },
          "targetDate": { "type": ["string", "null"], "format": "date" }
        },
        "minProperties": 1,
        "additionalProperties": false
      },
...
      "RecruitmentTaskResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/SuccessEnvelope" },
          { "properties": { "data": { "$ref": "#/components/schemas/RecruitmentTask" } } }
        ]
      },
      "RecruitmentTaskPageResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/SuccessEnvelope" },
          { "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/RecruitmentTask" } }, "meta": { "$ref": "#/components/schemas/PageMeta" } } }
        ]
      },
...
    }
  }
}